Provide mroe explciit instructions to explain the second settings button.
Both my wife and I found it difficult to understand what was happening
with the magical second settings button. In hindsight, it makes a lot of
sense and is a very clever way to avoid accidental misclicks by curious
children.

This extracts the text into a string to allow internationalisation, but
also (hopefully) clarifies the behaviour.