Move settings screen to activity, along with other UI refactorings.
Allows for a nice toolbar / material-design style. This means that
the exit button can be shown in a toolbar. It also means that to go back
from settings you can use a typical "Up" button in the toolbar.

Changes the forcing of landscape mode to the AndroidManifest rather
than programatically.

Only locks the screen in `onResume()`, not also `onCreate()`. Resume is
always called after create, so it accidentally calls it twice. Same for
`onRestart()` - the `onCreate()` is enough to lock at the right time.

Don't inflate the main activity view and then use it. Instead call
`setContentView(R.layout.*)` instead.

*Note:* The settings screen is not locked, so if a little one manages
to get to the settings screen, they can close the app. Given that the
app intentionally makes it dificult to get to settings, this seems like
a safe bet?