Seperate the melody from the playing of the melody.
I wasn't happy with the previous implementation where the static
information such as "These notes constitute these melodies"
was tied up with the state related to which note the melody was
currently up to.

This pulls the two things apart so the static notes always stay
static, and the "player" related logic is in a different class
hierarchy.