Unify soundset-related folder- and prefix-handling into SoundSet static helpers.
The old `SOUNDSET_DIR_PREFIX` already indicated in its name it wanted to live
in a SoundSet class.
Additionally, it's not just for DIRs, since it's also (ab?)used to get at the
translation keys.
As such, move the prefix to SoundSet, and isntead use SoundSet helpers
everywhere the prefix is currently used.

Also, the SoundSet class, which is already handling the raw folder access,
is a far better candidate for scanning that folder, than the current SettingsFragment.