Move 'envious' drawing logic from AppConfigTrigger to PianoCanvas.
Given that:
  1) all the drawing implementations it uses are already *in* PianoCanvas, and
  2) that it gets its required context *from* PianoCanvas, and
  3) it's the only usage of android context remaining in all of AppConfigTrigger,
see it through all the way:
Move the remaining drawing draw-logic to PianoCanvas as well.

To do this, we need to expose a bit more of AppConfigTrigger's state, which
actually brings us closer to a Model-View-Controller pattern, which is a sign we're
moving in the right direction.
AppConfigTrigger is now a proper (View-less) Model.
PianoCanvas still has a dual identity as View+Controller, but that's a problem
for another day.